Move renderers to their own file
see #64
